Added 2 test cases testing the response of the homepage
